Add fused modules to flash attn
